+++ src/orquestra/sdk/_base/_services.py
@@ -11,7 +11,7 @@ import subprocess
 from pathlib import Path
 from typing import Protocol
 
-from ._env import RAY_STORAGE_PATH_ENV, RAY_TEMP_PATH_ENV
+from ._env import RAY_PLASMA_PATH_ENV, RAY_STORAGE_PATH_ENV, RAY_TEMP_PATH_ENV
 
 ORQUESTRA_BASE_PATH = Path.home() / ".orquestra"
 
@@ -55,6 +55,16 @@ def ray_storage_path():
         return ORQUESTRA_BASE_PATH / "ray_storage"
 
 
+def ray_plasma_path():
+    """
+    See ``ray_temp_path``
+    """
+    try:
+        return Path(os.environ[RAY_PLASMA_PATH_ENV])
+    except KeyError:
+        return ORQUESTRA_BASE_PATH / "ray_plasma"
+
+
 # Timeout for inter-process commands (seconds).
 IPC_TIMEOUT = 20
 
@@ -87,6 +97,7 @@ class RayManager:
                 "--head",
                 f"--temp-dir={ray_temp_path()}",
                 f"--storage={ray_storage_path()}",
+                f"--plasma-directory={ray_plasma_path()}",
             ],
             check=False,
             timeout=IPC_TIMEOUT,